admin, server: improved `get_account_info` (#1423)
