+++ core/src/service/usage_service.rs
@@ -3,20 +3,10 @@ use serde::Serialize;
 use lockbook_shared::api::{FileUsage, GetUsageRequest, GetUsageResponse};
 use lockbook_shared::file_like::FileLike;
 use lockbook_shared::tree_like::{Stagable, TreeLike};
+use lockbook_shared::usage::bytes_to_human;
 
 use crate::{CoreError, RequestContext, Requester};
 use crate::{CoreResult, OneKey};
-
-const BYTE: u64 = 1;
-const KILOBYTE: u64 = BYTE * 1000;
-const MEGABYTE: u64 = KILOBYTE * 1000;
-const GIGABYTE: u64 = MEGABYTE * 1000;
-const TERABYTE: u64 = GIGABYTE * 1000;
-
-const KILOBYTE_MINUS_ONE: u64 = KILOBYTE - 1;
-const MEGABYTE_MINUS_ONE: u64 = MEGABYTE - 1;
-const GIGABYTE_MINUS_ONE: u64 = GIGABYTE - 1;
-const TERABYTE_MINUS_ONE: u64 = TERABYTE - 1;
 
 #[derive(Serialize, Debug)]
 pub struct UsageMetrics {
@@ -83,24 +73,4 @@ impl<Client: Requester> RequestContext<'_, '_, Client> {
         let readable = bytes_to_human(local_usage);
         Ok(UsageItemMetric { exact: local_usage, readable })
     }
-}
-
-pub fn bytes_to_human(size: u64) -> String {
-    let (unit, abbr) = match size {
-        0..=KILOBYTE_MINUS_ONE => (BYTE, "B"),
-        KILOBYTE..=MEGABYTE_MINUS_ONE => (KILOBYTE, "KB"),
-        MEGABYTE..=GIGABYTE_MINUS_ONE => (MEGABYTE, "MB"),
-        GIGABYTE..=TERABYTE_MINUS_ONE => (GIGABYTE, "GB"),
-        TERABYTE..=u64::MAX => (TERABYTE, "TB"),
-    };
-
-    let size_in_unit = size as f64 / unit as f64;
-    let dec = f64::trunc(size_in_unit.fract() * 100.0) / 100.0;
-
-    let num = format!("{:.2}", size_in_unit.trunc() + dec)
-        .trim_end_matches(['0'])
-        .trim_end_matches(['.'])
-        .to_owned();
-
-    format!("{} {}", num, abbr)
 }
